#  refactor: use b.Loop() in benchmark tests for better performance