# Description

BEFORE:
BenchmarkParseCoin-10 275319 4029 ns/op 2081 B/op 63 allocs/op
BenchmarkCoinsAdditionIntersect/A_1,B_1-10 5409805 297.8 ns/op 168 B/op 5 allocs/op
BenchmarkSig/secp256k1-10 8157 154376 ns/op 744 B/op 15 allocs/op
BenchmarkSig/secp256r1-10 18703 61438 ns/op 1296 B/op 24 allocs/op
BenchmarkIntValueRendererFormat-10 995870 1115 ns/op 784 B/op 19 allocs/op

AFTER:
BenchmarkParseCoin-10 270021 3902 ns/op 2081 B/op 63 allocs/op
BenchmarkCoinsAdditionIntersect/A_1,B_1-10 5444228 230.1 ns/op 168 B/op 5 allocs/op
BenchmarkSig/secp256k1-10 8332 147734 ns/op 744 B/op 15 allocs/op
BenchmarkSig/secp256r1-10 19350 60703 ns/op 1296 B/op 24 allocs/op
BenchmarkIntValueRendererFormat-10 1037956 1155 ns/op 784 B/op 19 allocs/op

reopen #25408